Follow the MySQL spec more closely when parsing statements
I've based this on the grammar layed out in
https://dev.mysql.com/doc/refman/5.7/en/create-table.html

This also defines language keywords in a way that makes writing the rest
of the grammer more succinctly, without caring about whitespace, and
using the !IdentifierStart trick that will prevent overly aggressive
matching, like:

    SELECT foo FROM bar;  // Matches
    SELECTfoo FROM bar;   // Parse error

This previously was not the case.

I'm also following the recommended casing and indenting style of the
author of the pegjs library:

- KEYWORD are ALL_CAPS
- RuleName in MixedCase
- variableName in camelCase

Rules that I have converted to the formal SQL grammer/spec, I have
changed the rule casing of.  So we can use that as a check list for
stuff that is still to-be-converted.