Refactor simulation functions into a class hierarchy
The class structure is as follows:

  - Database
      - Tables
          - Columns
          - Indexes
          - Foreign Keys
          - Primary Key

This matchs the previous structure, but these classes should now make it
easier to discover what methods exist on what objects.  In most cases,
you don't need to know about any of the Table or Column's methods, and
it's enough to simply call the method on the Database itself.

For example, to change a column's name, you'd still use:

   db2 = db.replaceColumn('my_table', 'old_col', newColumn);

And the Database will make sure this renaming consistently happens (i.e.
updating any references to it in other tables, etc) and returns a new DB
instance.